+++ src/main/java/br/jus/tjgo/litigantes/repository/RepositoryConfig.java
@@ -4,12 +4,12 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.data.rest.core.config.RepositoryRestConfiguration;
 import org.springframework.data.rest.webmvc.config.RepositoryRestConfigurerAdapter;
 
-import br.jus.tjgo.litigantes.model.Escritorio;
+import br.jus.tjgo.litigantes.model.Cliente;
 
 @Configuration
 public class RepositoryConfig extends RepositoryRestConfigurerAdapter {
     @Override
     public void configureRepositoryRestConfiguration(RepositoryRestConfiguration config) {
-        config.exposeIdsFor(Escritorio.class);
+        config.exposeIdsFor(Cliente.class);
     }
 }
